fix nlink_t madness
git-svn-id: svn+ssh://frodo/cluster/svn/projects/slash_nara@10026 3eda493b-6a19-0410-b2e0-ec8ea4dd8fda

+++ include/fidcache.h
@@ -234,11 +234,11 @@ extern struct psc_listcache	 fidcCleanList;
 #define DEBUG_STATBUF(stb, level)						\
 	psc_logs((level), PSS_GEN,						\
 	    "stb (%p) dev:%"PRIu64" inode:%"PRId64" mode:0%o "			\
-	    "nlink:%u uid:%u gid:%u "						\
+	    "nlink:%"PRIu64" uid:%u gid:%u "					\
 	    "rdev:%"PRIu64" sz:%"PRId64" blksz:%ld "				\
 	    "blkcnt:%"PRId64" atime:%lu mtime:%lu ctime:%lu",			\
 	    (stb), (uint64_t)(stb)->st_dev, (stb)->st_ino, (stb)->st_mode,	\
-	    (stb)->st_nlink, (stb)->st_uid, (stb)->st_gid,			\
+	    (uint64_t)(stb)->st_nlink, (stb)->st_uid, (stb)->st_gid,		\
 	    (uint64_t)(stb)->st_rdev, (stb)->st_size, (stb)->st_blksize,	\
 	    (stb)->st_blocks, (stb)->st_atime, (stb)->st_mtime,			\
 	    (stb)->st_ctime)
